+++ src/TavenemFrameworkSerializerContext.cs
@@ -1,0 +1,11 @@
+﻿using System.Text.Json.Serialization;
+using Tavenem.Blazor.Framework.Components.EmojiInput;
+
+namespace Tavenem.Blazor.Framework;
+
+[JsonSourceGenerationOptions(
+    PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase,
+    DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingDefault,
+    PropertyNameCaseInsensitive = true)]
+[JsonSerializable(typeof(Emoji[]))]
+internal partial class TavenemFrameworkSerializerContext : JsonSerializerContext;