+++ src/Utilities/EventHandlers.cs
@@ -6,6 +6,6 @@ namespace Tavenem.Blazor.Framework.Utilities;
 /// <summary>
 /// Event handlers for <c>Tavenem.Blazor.Framework</c>.
 /// </summary>
-[EventHandler("onfocuslost", typeof(EventArgs))]
+[EventHandler("onfocuslost", typeof(FocusLostEventArgs))]
 [EventHandler("dropdowntoggle", typeof(DropdownToggleEventArgs))]
-public static class EventHandlers { }
+public static class EventHandlers;